feat: service

+++ src/lib/runner.ts
@@ -1,6 +1,12 @@
 import { fork, Serializable } from 'child_process'
 import { Logger } from '@src/lib/logger'
-import { TRunnerResult, RUNNER_STATUS } from '@type/runner.type'
+import { eventEmitter } from '@src/lib/eventEmitter'
+import {
+  RUNNER_IPC_TYPE,
+  TRunnerIpcMessage,
+  TRunnerResult,
+  RUNNER_STATUS,
+} from '@type/runner.type'
 
 export function run(
   name: string,
@@ -35,19 +41,28 @@ export function run(
     }, timeout)
 
     subprocess.once('spawn', () => {
-      subprocess.send(payload)
+      subprocess.send({
+        type: RUNNER_IPC_TYPE.START,
+        data: payload,
+      } as TRunnerIpcMessage)
     })
 
-    subprocess.once('message', (message) => {
+    subprocess.on('message', (message: TRunnerIpcMessage) => {
       if (handled) return
-      handled = true
-      clearTimeout(timer)
-      subprocess.removeAllListeners()
-      subprocess.kill()
-      resolve({
-        status: RUNNER_STATUS.Success,
-        result: message,
-      })
+      if (message.type === RUNNER_IPC_TYPE.END) {
+        handled = true
+        clearTimeout(timer)
+        subprocess.removeAllListeners()
+        subprocess.kill()
+        resolve({
+          status: RUNNER_STATUS.Success,
+          result: message.data,
+        })
+      } else if (message.type === RUNNER_IPC_TYPE.SERVICE) {
+        if (message.name) {
+          eventEmitter.emit(message.name, subprocess, message.data, message.id)
+        }
+      }
     })
 
     subprocess.once('exit', (code) => {
